backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pyyaml==6.0.2
requests==2.32.3
pandas==2.2.3
openpyxl==3.1.5
python-multipart==0.0.12
